# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/orders": {
            "get": {
                "summary": "findAllOrders",
                "description": "Display a listing of the resource.",
                "operationId": "8f3183bfa0ecb8a2c864cbf972b8061d",
                "responses": {
                    "200": {
                        "description": "A list of orders"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            },
            "post": {
                "summary": "Save order",
                "operationId": "b6490f9d9026dbe871b41207e118fffb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "cur",
                        "in": "path",
                        "description": "Parameter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store order"
                    },
                    "default": {
                        "description": "no products"
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "summary": "Show a order",
                "description": "Display the specified resource.",
                "operationId": "79c1e32fc3603170dfd7659562133aeb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter with mutliple examples",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "summary": "Register a new user",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "Password of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Register user"
                    },
                    "default": {
                        "description": "key"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "summary": "Login user",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "Password of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login user"
                    },
                    "default": {
                        "description": "key"
                    }
                }
            }
        },
        "/api/resource.json": {
            "get": {
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}